More LH stuff (#9)
* Change the way ModelFactory exports and formats to o2r

* Correct geolayout and modelDL writes

* Correct more exports based on decomp expectations

* Allow RelWithDebInfo

* Rework the way Torch exports maps to be consistent with other ports

* Bake asset symbolmap into the o2r as aBKAssetTable for fast lookups that don't rely on a present yaml file

* Fix raw textures from models and capture sprite dimensions

* Correct writes for dialog and sprites

* Fix animation bitfield extraction

+++ src/factories/bk64/AnimFactory.cpp
@@ -58,9 +58,9 @@ ExportResult BK64::AnimBinaryExporter::Export(std::ostream& write, std::shared_p
         writer.Write(file.mTransformType);
         writer.Write((uint32_t)file.mData.size());
         for (const auto& fileData : file.mData) {
-            writer.Write(fileData.unk0_15);
-            writer.Write(fileData.unk0_14);
-            writer.Write(fileData.unk0_13);
+            writer.Write(static_cast<uint8_t>(fileData.unk0_15));
+            writer.Write(static_cast<uint8_t>(fileData.unk0_14));
+            writer.Write(static_cast<uint16_t>(fileData.unk0_13));
             writer.Write(fileData.unk2);
         }
     }
